Move tests outside of package to top level directory
Cleanly separates tests from the package itself. Prevents the tests
being installed to site-packages.Tests are still distributed with the
source distribution.
