+++ doc/source/conf.py
@@ -71,7 +71,7 @@ release = version_info.release_string()
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
-exclude_patterns = ['_build', 'api/setup.rst', 'api/requests_mock.tests.rst']
+exclude_patterns = ['_build', 'api/setup.rst', 'api/tests.rst']
 
 # The reST default role (used for this markup: `text`) to use for all documents
 #default_role = None
