+++ setup.cfg
@@ -19,7 +19,7 @@ classifier =
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3.4
     Topic :: Software Development :: Testing
-test_suite = requests_mock.tests
+test_suite = tests
 
 [files]
 packages = requests_mock
