+++ tests/test_fixture.py
@@ -13,7 +13,7 @@
 import requests
 import requests_mock
 from requests_mock.contrib import fixture
-from requests_mock.tests import base
+from tests import base
 
 
 class MockingTests(base.TestCase):
