+++ tox.ini
@@ -11,7 +11,7 @@ deps =
     pbr
 
 [testenv:pep8]
-commands = flake8 requests_mock
+commands = flake8 requests_mock tests
 deps =
     flake8
     {[testenv]deps}
